requirements.txt
anthropic==0.75.0
build==1.3.0
google-genai==1.56.0
google-generativeai==0.8.6
grpcio-status==1.71.2
importlib-metadata==8.0.0
openai==2.13.0
pip-chill==1.0.3
python-dotenv==1.2.1
streamlit==1.52.2
tomli==2.0.1
datasets>=2.14.0
jaraco.collections==5.1.0
platformdirs==4.2.2
plotly>=5.15.0

